feat: brt-service add db migrations

+++ brt-service/src/main/java/ru/ellio/brtservice/model/Client.java
@@ -23,7 +23,7 @@ public class Client {
     @JoinColumn(name = "tariff_id")
     private Tariff tariff;
 
-    private long balance;
+    private double balance;
 
     @OneToMany
     private List<Operation> payload;
