+++ brt-service/src/main/java/ru/ellio/brtservice/model/Operation.java
@@ -26,5 +26,5 @@ public class Operation {
 
     private LocalTime duration;
 
-    private long cost;
+    private double cost;
 }
